move common items to validation/ root
